main service  Events getEvents Получение подробной информации об опубликованном событии по его идентификатору

+++ stat/stat-dto/src/main/java/dto/ViewStatsDTO.java
@@ -10,5 +10,5 @@ import lombok.*;
 public class ViewStatsDTO {
     private String app; // Название сервиса
     private String uri; // URI сервиса
-    private Long hits; // Количество просмотров
+    private Integer hits; // Количество просмотров
 }
